+++ R/ggscreeplot.r
@@ -3,12 +3,16 @@
 #' Produces scree plots (Cattell, 1966) of the variance proportions explained by each dimension against dimension number from 
 #' various dimension reduction techniques
 #'
-#' @param pcobj   an object representing a linear dimension technique, such a returned by \code{\link[stats]{prcomp}} 
-#'                or \code{\link[stats]{princomp}} or \code{\link[FactoMineR]{PCA}} or \code{\link[MASS]{lda}}
+#' @param pcobj   an object returned by \code{\link[stats]{prcomp}}, \code{\link[stats]{princomp}}, 
+#'                \code{\link[FactoMineR]{PCA}}, \code{\link[ade4]{dudi.pca}}, or \code{\link[MASS]{lda}}
 #' @param type    the type of scree plot.  
 #'                'pev' corresponds proportion of explained variance, i.e. the eigenvalues divided by the trace. 
 #'                'cev' corresponds to the cumulative proportion of explained variance, i.e. the partial sum of the first k eigenvalues divided by the trace.
 #' @param size    point size
+#' @param shape   shape of the points
+#' @param color   color for points and line
+#' @param linetype type of line
+#' @param linewidth width of line
 #' 
 #' @returns A ggplot2 object
 #' @export
@@ -21,7 +25,11 @@
 #'
 ggscreeplot <- function(pcobj, 
                         type = c('pev', 'cev'),
-                        size = 4) 
+                        size = 4,
+                        shape = 1,
+                        color = "black",
+                        linetype = 1,
+                        linewidth = 1) 
 {
   # Recover the SVD
   if(inherits(pcobj, 'prcomp')){
@@ -35,16 +43,19 @@ ggscreeplot <- function(pcobj,
     d <- unlist(sqrt(pcobj$eig)[1])
   } else if(inherits(pcobj, "lda")) {
     d <- pcobj$svd
-  } else {
-    stop('Expected a object of class "prcomp", "princomp", "PCA", or "lda"')
+  } else if(inherits(pcobj, 'pca') & inherits(pcobj, 'dudi')){
+    d <- sqrt(pcobj$eig)
+  }
+  else {
+    stop('Expected a object of class "prcomp", "princomp", "PCA", c("pca", "dudi") or "lda"')
   }
   
 
   type <- match.arg(type)
-  d <- pcobj$sdev^2
+  dsq <- d^2
   yvar <- switch(type, 
-                 pev = d / sum(d), 
-                 cev = cumsum(d) / sum(d))
+                 pev = dsq / sum(dsq), 
+                 cev = cumsum(dsq) / sum(dsq))
 
   yvar.lab <- switch(type,
                      pev = 'Proportion of explained variance',
@@ -56,5 +67,6 @@ ggscreeplot <- function(pcobj,
   ggplot(data = df, aes(x = PC, y = yvar)) + 
     xlab('Principal component number') + 
     ylab(yvar.lab) +
-    geom_point(size = size) + geom_path()
+    geom_point(size = size, shape = shape, color = color) + 
+    geom_path(linetype = linetype, linewidth = linewidth, color = color)
 }